requirements.txt
torch==2.5.1
numpy==2.1.3
openai-whisper==20240930
python-dotenv==1.0.0
PyAudio==0.2.14
Flask==3.0.2
Werkzeug==3.0.1